# Optimized a bit ebrisk